Explore other sensors
"GRIP" sensor seems to activate when I grab the top
of the device by the sides. I've only ever seen the following two
possibities for values:
  [5, 0, 0]
  [0, 0, 0]

I have not figured out how to trigger one or the other

"MOTION" sensor *only* seems to fire when I'm setting down my
device and *not* trying to trigger it. Here are the values I've
seen:

2019-09-09 09:59:52.273 7351-7351/com.example.scroggo.sensorpractice D/SCROGGO: Saw some motion: [86.0, -1.6112977, 11.801021]
2019-09-09 10:01:20.088 7351-7351/com.example.scroggo.sensorpractice D/SCROGGO: Saw some motion: [86.0, -2.4492683, 7.6722193]
2019-09-09 10:01:29.014 7351-7351/com.example.scroggo.sensorpractice D/SCROGGO: Saw some motion: [86.0, -2.8107927, 9.697714]
2019-09-09 10:01:39.517 7351-7351/com.example.scroggo.sensorpractice D/SCROGGO: Saw some motion: [86.0, -2.955642, 9.024943]
2019-09-09 10:01:46.774 7351-7351/com.example.scroggo.sensorpractice D/SCROGGO: Saw some motion: [86.0, -0.25617957, 9.384073]

Note that the first value is always 86. I don't know what these represent.